@@ -4,38 +4,28 @@
     <!-- PROJECT METADATA -->
     <groupId>ch.qos.logback</groupId>
     <artifactId>logback-s3-rolling-policy</artifactId>
-    <version>GIT-SNAPSHOT</version>
+    <version>0.0.1</version>
+    <packaging>jar</packaging>
 
-    <!-- CODE MANAGEMENT -->
-    <scm>
-        <tag>HEAD</tag>
-        <url>https://git@github.com/link-nv/logback-s3</url>
-        <connection>scm:git:ssh://git@github.com/link-nv/logback-s3</connection>
-        <developerConnection>scm:git:ssh://git@github.com/link-nv/logback-s3</developerConnection>
-    </scm>
-
-    <reporting>
-        <outputDirectory>target/site</outputDirectory>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-javadoc-plugin</artifactId>
-                <version>${maven-javadoc-plugin.version}</version>
-            </plugin>
-        </plugins>
-    </reporting>
+    <name>logback-s3-rolling-policy</name>
+    <url>https://github.com/komuro-hiraku/logback-s3-rolling-policy</url>
 
     <properties>
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <java.version>1.8</java.version>
 
+        <!-- 認証情報 -->
+        <github.global.server>github</github.global.server>
+
+        <!-- Version情報 -->
         <maven-javadoc-plugin.version>2.8.1</maven-javadoc-plugin.version>
         <maven-surefire-report-plugin.version>2.9</maven-surefire-report-plugin.version>
         <maven-clean-plugin.version>2.4.1</maven-clean-plugin.version>
         <maven-resources-plugin.version>2.5</maven-resources-plugin.version>
         <maven-release-plugin.version>2.1</maven-release-plugin.version>
         <maven-install-plugin.version>2.3.1</maven-install-plugin.version>
-        <maven-compiler-plugin.version>2.0.2</maven-compiler-plugin.version>
+        <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
         <maven-surefire-plugin.version>2.9</maven-surefire-plugin.version>
         <maven-surefire-report-plugin.version>2.5</maven-surefire-report-plugin.version>
         <maven-assembly-plugin.version>2.2-beta-5</maven-assembly-plugin.version>
